add code method and tests
